LLM Triton/app/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pyngrok
